Script to separate amalgamated genes
This script reads the gene names in the second column and does the following: splits genes into seperate rows if they happen to occur in the same raw sepearated by (,) and assigns the same expression values per each sepearated gene (as the same XLOC ids) while the rest of raws are just outputed. The programme ignores newly discovered genes. This code does NOT identify genes that occur multiple times in the dataset.